Handle response errors in api requests
We sometimes have the case that the prunner doesn't send proper responses and the `json_decode` method returns null instead of an array which leads to errors in the `PipelinesAndJobsResponse::fromArray`.
When the response is first checked and a runtime exception is thrown this can be handled for the user.